Replace sequential line loader and split declarations and implementation
