+++ figures-counter/CMakeLists.txt
@@ -6,11 +6,12 @@ set (PROJECT_EXECUTABLE ${PROJECT_NAME})
 # Compile to a static library
 add_library (${PROJECT_LIB} STATIC 
 	impl/figures-counter.cpp
-	impl/error.cpp
-	"impl/internal/line_loader.cpp"
-	"impl/internal/sequential_figures_counter.cpp"
+	impl/error.cpp	
 	"impl/internal/bmp_figures_counter.cpp"
-)
+	"impl/internal/bmp_line_loader.cpp"
+	"impl/internal/bmp.cpp"
+	"impl/internal/disjoint_set.cpp"
+ "impl/includes/bmp.h")
 target_include_directories(${PROJECT_LIB} PUBLIC	
 	${PROJECT_SOURCE_DIR}/includes
 )
@@ -21,7 +22,7 @@ target_include_directories(${PROJECT_LIB} PRIVATE
 # Compile to an executable
 add_executable (${PROJECT_EXECUTABLE}
 	main.cpp
-)
+ "impl/includes/bmp.h")
 target_link_libraries (${PROJECT_EXECUTABLE}
 	${PROJECT_LIB}
 )
