+++ figures-counter/includes/figures-counter.h
@@ -48,8 +48,6 @@ bool operator==(char lhs, Cell rhs) noexcept;
  * @throws error::bad_input If any of the input parameters is not in the described format
  */
 
-size_t count_figures(size_t rows, size_t cols, std::istream& data);
-
 size_t count_figures(const std::filesystem::path& bmp);
 
 } // namespace figures_counter
